+++ CHANGELOG.md
@@ -3,9 +3,9 @@
 ## 0.1.2
 
 ### Improvements
+- Auto-complete fields from referenced temp-table
 - Broadcast changes to referenced documents
 - Request file name to save map file
-- Display all fields from temp-table "like"
 
 ## 0.1.1